--- snmpsim/mltsplit.py
@@ -1,9 +1,0 @@
-#
-# This file is part of snmpsim software.
-#
-# Copyright (c) 2010-2019, Ilya Etingof <etingof@gmail.com>
-# License: http://snmplabs.com/snmpsim/license.html
-#
-# Like string.split but first tries to use composite separator as an
-# escaping aid
-
